Restructure front end pages
